+++ context-tutorial/src/components/SelectColors.js
@@ -1,0 +1,41 @@
+import React from "react";
+import { ColorConsumer } from "../contexts/color";
+
+const colors = ["red", "orange", "yellow", "green", "blue", "indigo", "violet"];
+
+const SelectColors = () => {
+  return (
+    <div>
+      <h2>색상을 선택하세요.</h2>
+      <ColorConsumer>
+        {({ actions }) => (
+          <div style={{ display: "flex" }}>
+            {colors.map((color) => (
+              <div
+                key={color}
+                style={{
+                  background: color,
+                  width: 24,
+                  height: 24,
+                  cursor: "pointer",
+                }}
+                onClick={() => {
+                  actions.setColor(color);
+                  console.log("왼쪽 클릭 중...");
+                }}
+                onContextMenu={(e) => {
+                  e.preventDefault(); // 오른쪽 버튼 클릭 시 메뉴가 뜨는 것을 무시
+                  actions.setSubColor(color);
+                  console.log("오른쪽 클릭 중...");
+                }}
+              />
+            ))}
+          </div>
+        )}
+      </ColorConsumer>
+      <hr />
+    </div>
+  );
+};
+
+export default SelectColors;
